requirements.txt
# --- Мережевий стек та API (Критично для Binance/Bybit) ---
ccxt==4.4.22
aiodns==3.6.1
pycares==4.11.0


# --- Обробка даних (Розрахунок спреду та OHLCV) ---
pandas==2.3.3
#numpy==2.2.6
#ujson==5.12.0

# --- Графічний інтерфейс та Backend ---
pywebview==6.1

pydantic==2.12.5
#uvicorn==0.32.1


